Add new plugin SK Password Resetter
Used to reset the password for all users except the one called
"wpadmin". This is needed because of a change done to the SSO-plugin
which makes sure that all passwords are being checked by the username in
lowercase. Because of this we need to make sure all passwords in the DB
is store based on these conditions.